MP-151 :truck: Rename: JWT 파일 이동
- 패키지 구조 변경에 따라 기존 파일 이동
