+++ src/main/java/kr/modusplant/modules/jwt/app/error/TokenKeyCreationException.java
@@ -1,4 +1,4 @@
-package kr.modusplant.modules.jwt.error;
+package kr.modusplant.modules.jwt.app.error;
 
 public class TokenKeyCreationException extends RuntimeException {
     public TokenKeyCreationException(String message, Throwable cause) {
