+++ src/main/java/kr/modusplant/modules/jwt/app/service/TokenProvider.java
@@ -1,9 +1,9 @@
-package kr.modusplant.modules.jwt.domain.service;
+package kr.modusplant.modules.jwt.app.service;
 
 import io.jsonwebtoken.*;
 import jakarta.annotation.PostConstruct;
-import kr.modusplant.modules.jwt.error.InvalidTokenException;
-import kr.modusplant.modules.jwt.error.TokenKeyCreationException;
+import kr.modusplant.modules.jwt.app.error.InvalidTokenException;
+import kr.modusplant.modules.jwt.app.error.TokenKeyCreationException;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
